apps/demo3_dash_dbsql/requirements.txt
dash
databricks-sdk==0.33.0
databricks-sql-connector
pandas==2.2.3
